+++ src/main/java/act/cli/CsvView.java
@@ -22,10 +22,7 @@ package act.cli;
 
 import act.util.PropertySpec;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * **Note** this is deprecated, please use `act.util.CsvView` instead
@@ -54,5 +51,6 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.CLASS)
 @Target(ElementType.METHOD)
 @Deprecated
+@Inherited
 public @interface CsvView {
 }