+++ src/main/java/act/cli/TreeView.java
@@ -23,10 +23,7 @@ package act.cli;
 import act.cli.view.CliView;
 import act.util.PropertySpec;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * Mark a command method return value shall be displayed as tree structure. e.g
@@ -39,5 +36,6 @@ import java.lang.annotation.Target;
  */
 @Retention(RetentionPolicy.CLASS)
 @Target(ElementType.METHOD)
+@Inherited
 public @interface TreeView {
 }