+++ src/main/java/act/util/CsvView.java
@@ -22,10 +22,7 @@ package act.util;
 
 import act.cli.TableView;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * Mark a command method return value shall be displayed using a csv file format. e.g
@@ -49,5 +46,6 @@ import java.lang.annotation.Target;
  */
 @Retention(RetentionPolicy.CLASS)
 @Target(ElementType.METHOD)
+@Inherited
 public @interface CsvView {
 }