+++ src/main/java/act/util/JsonView.java
@@ -22,10 +22,7 @@ package act.util;
 
 import act.cli.TableView;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * Mark a command method return value shall be displayed as a JSON object. e.g
@@ -56,5 +53,6 @@ import java.lang.annotation.Target;
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD, ElementType.TYPE})
+@Inherited
 public @interface JsonView {
 }